macropad/bridge/requirements.txt
pyserial>=3.5
spotipy>=2.24.0
Pillow>=10.0.0
pyinstaller>=6.0.0
